Change in how the renderer detects list data
Previously, the renderer would check whether `view.action` was `'list'`,
and if not it would assume the data was a single resource instance.
This can cause problems when a custom route returns list data.

With this commit, the renderer instead checks whether the serializer
that produced that data has `many` set to True.