Fix/auto prefetch with m2m (#333)
* Get the field queryset in a more reliable way

That works through many to many field descriptors, previous way only
worked through many to ones.

* Update test views to use optimised viewset

* Rename comment_set to comments, use json viewset

Updates the tests to use the optimised queryset views provided in
views.py. For these to work, the serializer fields need to match the
model names so I have just changed the related_name in the test models.

This could be improved elsewhere, with the queryset prefetching based on
the serializer. But for the most basic use case, this is fine.

* Correctly prefetch both forward and reverse

Based on the relation descriptor.

* Support field rename in older django

* Update to py2 friendly super

* Replace hasattr with explicit version check